--- PublishWorkflow/PublishTimePeriodWriteInterface.php
@@ -1,31 +1,0 @@
-<?php
-
-namespace Symfony\Cmf\Bundle\CoreBundle\PublishWorkflow;
-
-/**
- * Interface to expose editable time period publishing information.
- */
-interface PublishTimePeriodWriteInterface extends PublishTimePeriodInterface
-{
-    /**
-     * Set the date from which the content should
-     * be considered publishable.
-     *
-     * Setting a NULL value asserts that the content
-     * has always been publishable.
-     *
-     * @param \DateTime|null $publishDate
-     */
-    public function setPublishStartDate(\DateTime $publishDate = null);
-
-    /**
-     * Set the date at which the content should
-     * stop being published.
-     *
-     * Setting a NULL value asserts that the
-     * content will always be publishable.
-     *
-     * @param \DateTime|null $publishDate
-     */
-    public function setPublishEndDate(\DateTime $publishDate = null);
-}
